build: align Exposed Timefold persistence with Solver 2.x
Move the Exposed Timefold persistence module to the actual Timefold Solver 2.1 API and materialize the related central catalog upgrades.

Constraint: Timefold Solver 2.1 removed *LongScore classes and no longer publishes timefold-solver-test or persistence-common artifacts.

Rejected: keep removed score helpers | they cannot compile against the 2.1 public API.

Confidence: high

Scope-risk: broad

Directive: Do not reintroduce Timefold 1.x score package imports or removed 1.x artifacts.

Tested: ./gradlew :bluetape4k-exposed-timefold-solver-persistence:compileTestKotlin --no-daemon; central symlink workspace sync check.

Not-tested: Full repository test suite after AWS/Fory catalog materialization.

+++ exposed/exposed-timefold-solver-persistence/src/test/kotlin/io/bluetape4k/timefold/solver/exposed/api/score/buildin/HardSoftScoreTest.kt
@@ -1,6 +1,6 @@
 package io.bluetape4k.timefold.solver.exposed.api.score.buildin
 
-import ai.timefold.solver.core.api.score.buildin.hardsoft.HardSoftScore
+import ai.timefold.solver.core.api.score.HardSoftScore
 import io.bluetape4k.exposed.dao.entityToStringBuilder
 import io.bluetape4k.exposed.dao.idEquals
 import io.bluetape4k.exposed.dao.idHashCode
@@ -50,8 +50,8 @@ class HardSoftScoreTest: AbstractScoreExposedTest() {
 
             val name = faker.name().name()
             val hardSoftScore = HardSoftScore.of(
-                faker.random().nextInt(),
-                faker.random().nextInt()
+                faker.random().nextLong(),
+                faker.random().nextLong()
             )
 
             val id = T1.insertAndGetId {
@@ -78,8 +78,8 @@ class HardSoftScoreTest: AbstractScoreExposedTest() {
             val saved = E1.new {
                 this.name = faker.name().name()
                 this.hardSoftScore = HardSoftScore.of(
-                    faker.random().nextInt(0, 100),
-                    faker.random().nextInt(1, 100)
+                    faker.random().nextLong(0, 100),
+                    faker.random().nextLong(1, 100)
                 )
             }
             log.debug { "saved=$saved" }
